backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
